Issue #129 Change the latitude and longitude data types from decimal to float

+++ locations/models.py
@@ -5,8 +5,8 @@ from organizations.models import Organization
 class Location(TranslatableModel):
     organization = models.ForeignKey(Organization, on_delete=models.CASCADE)
     name = models.CharField(max_length=200)
-    latitude = models.DecimalField(blank=True, null=True, max_digits=9, decimal_places=6)
-    longitude = models.DecimalField(blank=True, null=True, max_digits=9, decimal_places=6)
+    latitude = models.FloatField(blank=True, null=True)
+    longitude = models.FloatField(blank=True, null=True)
     translations = TranslatedFields(
         description=models.TextField(blank=True)
     )
